move some methods that operate on a Job into the Job class
